+++ episode.go
@@ -1,0 +1,39 @@
+package tvdb
+
+// Episode represents a TV show episode on TheTVDB.
+type Episode struct {
+	ID                    uint64   `xml:"id"`
+	CombinedEpisodeNumber string   `xml:"Combined_episodenumber"`
+	CombinedSeason        uint64   `xml:"Combined_season"`
+	DvdChapter            string   `xml:"DVD_chapter"`
+	DvdDiscID             string   `xml:"DVD_discid"`
+	DvdEpisodeNumber      string   `xml:"DVD_episodenumber"`
+	DvdSeason             string   `xml:"DVD_season"`
+	Director              pipeList `xml:"Director"`
+	EpImgFlag             string   `xml:"EpImgFlag"`
+	EpisodeName           string   `xml:"EpisodeName"`
+	EpisodeNumber         uint64   `xml:"EpisodeNumber"`
+	FirstAired            string   `xml:"FirstAired"`
+	GuestStars            string   `xml:"GuestStars"`
+	ImdbID                string   `xml:"IMDB_ID"`
+	Language              string   `xml:"Language"`
+	Overview              string   `xml:"Overview"`
+	ProductionCode        string   `xml:"ProductionCode"`
+	Rating                string   `xml:"Rating"`
+	RatingCount           string   `xml:"RatingCount"`
+	SeasonNumber          uint64   `xml:"SeasonNumber"`
+	Writer                pipeList `xml:"Writer"`
+	AbsoluteNumber        string   `xml:"absolute_number"`
+	Filename              string   `xml:"filename"`
+	LastUpdated           string   `xml:"lastupdated"`
+	SeasonID              uint64   `xml:"seasonid"`
+	SeriesID              uint64   `xml:"seriesid"`
+	ThumbAdded            string   `xml:"thumb_added"`
+	ThumbHeight           string   `xml:"thumb_height"`
+	ThumbWidth            string   `xml:"thumb_width"`
+}
+
+// EpisodeList represents a list of TV show episodes.
+type EpisodeList struct {
+	Episodes []*Episode `xml:"Episode"`
+}
