+++ tvdb.go
@@ -5,12 +5,9 @@ import (
 	"encoding/json"
 	"fmt"
 	"bytes"
-	"time"
 	"io/ioutil"
 	"net/http"
 	"regexp"
-	"strconv"
-	"strings"
 )
 
 const (
@@ -70,30 +67,6 @@ const (
 // SearchSeriesRegex is used for series web search matching.
 var SearchSeriesRegex = regexp.MustCompile(SearchSeriesRegexPattern)
 
-// PipeList type representing pipe-separated string values.
-type PipeList []string
-
-// UnmarshalXML unmarshals an XML element with string value into a pip-separated list of strings.
-func (pipeList *PipeList) UnmarshalJSON(data []byte) (err error) {
-	*pipeList = strings.Split(strings.Trim(string(data), "|"), "|")
-
-	return
-}
-
-type unixTime time.Time
-
-func (t *unixTime) UnmarshalJSON(data []byte) (err error) {
-	unixSeconds, err := strconv.ParseInt(string(data), 10, 64)
-
-	if err != nil {
-		return
-	}
-
-	*t = unixTime(time.Unix(unixSeconds, 0))
-
-	return
-}
-
 type TheTVDB struct {
 	apiKey string
 	jwt jwt
@@ -113,7 +86,7 @@ type Actor struct {
 	ID uint64 `json:"id"`
 	SeriesID uint64 `json:"seriesId"`
 	Name string `json:"name"`
-	Role PipeList `json:"role"`
+	Role pipeList `json:"role"`
 	SortOrder uint64 `json:"sortOrder"`
 	Image string `json:"image"`
 	ImageAuthor uint64 `json:"imageAuthor"`
@@ -244,95 +217,6 @@ func (tvdb *TheTVDB) RefreshToken() (err error) {
 
 	return
 }
-
-// Episode represents a TV show episode on TheTVDB.
-type Episode struct {
-	ID                    uint64   `xml:"id"`
-	CombinedEpisodeNumber string   `xml:"Combined_episodenumber"`
-	CombinedSeason        uint64   `xml:"Combined_season"`
-	DvdChapter            string   `xml:"DVD_chapter"`
-	DvdDiscID             string   `xml:"DVD_discid"`
-	DvdEpisodeNumber      string   `xml:"DVD_episodenumber"`
-	DvdSeason             string   `xml:"DVD_season"`
-	Director              PipeList `xml:"Director"`
-	EpImgFlag             string   `xml:"EpImgFlag"`
-	EpisodeName           string   `xml:"EpisodeName"`
-	EpisodeNumber         uint64   `xml:"EpisodeNumber"`
-	FirstAired            string   `xml:"FirstAired"`
-	GuestStars            string   `xml:"GuestStars"`
-	ImdbID                string   `xml:"IMDB_ID"`
-	Language              string   `xml:"Language"`
-	Overview              string   `xml:"Overview"`
-	ProductionCode        string   `xml:"ProductionCode"`
-	Rating                string   `xml:"Rating"`
-	RatingCount           string   `xml:"RatingCount"`
-	SeasonNumber          uint64   `xml:"SeasonNumber"`
-	Writer                PipeList `xml:"Writer"`
-	AbsoluteNumber        string   `xml:"absolute_number"`
-	Filename              string   `xml:"filename"`
-	LastUpdated           string   `xml:"lastupdated"`
-	SeasonID              uint64   `xml:"seasonid"`
-	SeriesID              uint64   `xml:"seriesid"`
-	ThumbAdded            string   `xml:"thumb_added"`
-	ThumbHeight           string   `xml:"thumb_height"`
-	ThumbWidth            string   `xml:"thumb_width"`
-}
-
-// EpisodeList represents a list of TV show episodes.
-type EpisodeList struct {
-	Episodes []*Episode `xml:"Episode"`
-}
-
-/*// GetDetail gets more detail for all TV shows in a list.
-func (seriesList *SeriesList) GetDetail() (err error) {
-	for seriesIndex := range seriesList.Series {
-		if err = seriesList.Series[seriesIndex].GetDetail(); err != nil {
-			return
-		}
-	}
-
-	return
-}
-
-// GetDetail gets more detail for a TV show, including information on it's episodes.
-func (series *Series) GetDetail() (err error) {
-	response, err := http.Get(fmt.Sprintf(GetDetailURL, APIKey, strconv.FormatUint(series.ID, 10)))
-
-	if err != nil {
-		return
-	}
-
-	data, err := ioutil.ReadAll(response.Body)
-
-	if err != nil {
-		return
-	}
-
-	// Unmarshal data into SeriesList object
-	seriesList := SeriesList{}
-	if err = xml.Unmarshal(data, &seriesList); err != nil {
-		return
-	}
-
-	// Copy the first result into the series object
-	*series = *seriesList.Series[0]
-
-	episodeList := EpisodeList{}
-
-	if err = xml.Unmarshal(data, &episodeList); err != nil {
-		return
-	}
-
-	if series.Seasons == nil {
-		series.Seasons = make(map[uint64][]*Episode)
-	}
-
-	for _, episode := range episodeList.Episodes {
-		series.Seasons[episode.SeasonNumber] = append(series.Seasons[episode.SeasonNumber], episode)
-	}
-
-	return
-}*/
 
 func (tvdb *TheTVDB) Languages() (languages []Language, err error) {
 	// Check JWT expiry.
@@ -546,108 +430,3 @@ func (tvdb *TheTVDB) GetSeriesByID(id uint64) (series Series, err error) {
 
 	return
 }
-
-/*// GetSeries gets a list of TV series by name, by performing a simple search.
-func (tvdb *TheTVDB) GetSeries(name string) (seriesList SeriesList, err error) {
-	response, err := http.Get(fmt.Sprintf(GetSeriesURL, url.QueryEscape(name)))
-
-	if err != nil {
-		return
-	}
-
-	data, err := ioutil.ReadAll(response.Body)
-
-	if err != nil {
-		return
-	}
-
-	err = xml.Unmarshal(data, &seriesList)
-
-	return
-}
-
-// GetSeriesByIMDBID gets series from IMDb's ID.
-func GetSeriesByIMDBID(id string) (series *Series, err error) {
-	response, err := http.Get(fmt.Sprintf(GetSeriesByIMDBIDURL, id))
-
-	if err != nil {
-		return
-	}
-
-	data, err := ioutil.ReadAll(response.Body)
-
-	if err != nil {
-		return
-	}
-
-	seriesList := SeriesList{}
-
-	if err = xml.Unmarshal(data, &seriesList); err != nil {
-		return
-	}
-
-	if len(seriesList.Series) != 1 {
-		err = errors.New("incorrect number of series")
-
-		return
-	}
-
-	series = seriesList.Series[0]
-
-	return
-}
-
-// SearchSeries searches for TV shows by name, using the more sophisticated
-// search on TheTVDB's homepage. This is the recommended search method.
-func SearchSeries(name string, maxResults int) (seriesList SeriesList, err error) {
-	response, err := http.Get(fmt.Sprintf(SearchSeriesURL, url.QueryEscape(name)))
-
-	if err != nil {
-		return
-	}
-
-	buf, err := ioutil.ReadAll(response.Body)
-
-	if err != nil {
-		return
-	}
-
-	groups := SearchSeriesRegex.FindAllSubmatch(buf, -1)
-	doneSeriesIDs := make(map[uint64]struct{})
-
-	for _, group := range groups {
-		seriesID := uint64(0)
-		var series *Series
-		seriesID, err = strconv.ParseUint(string(group[2]), 10, 64)
-
-		if _, ok := doneSeriesIDs[seriesID]; ok {
-			continue
-		}
-
-		if err != nil {
-			return
-		}
-
-		series, err = GetSeriesByID(seriesID)
-
-		if err != nil {
-			// Some series can't be found, so we will ignore these.
-			if _, ok := err.(*xml.SyntaxError); ok {
-				err = nil
-
-				continue
-			} else {
-				return
-			}
-		}
-
-		seriesList.Series = append(seriesList.Series, series)
-		doneSeriesIDs[seriesID] = struct{}{}
-
-		if len(seriesList.Series) == maxResults {
-			break
-		}
-	}
-
-	return
-}*/
